Add raw data support
This commit adds the ability to return the raw topic and message buffer,
allowing easier handling of binary data from mqtt inside the graphql endpoint
itself.

Enable by setting the `rawData` option. If it is not set, it behaves
identically to the existing JSON/String path.

It also upgrades Mocha to 7.0.1, to resolve multiple security
vulnerabilities with Mocha.